add optional async extension for lsp clients
The new crate lsp_client holds functions that can be used to apply the
features from ng_async. This way the json parsing is performed in a
rust thread.

There's currently only support for lsp-mode.
