+++ rust_src/crates/lsp_client/.gitignore
@@ -1,0 +1,1 @@
+out/
