+++ rust_src/crates/lsp_client/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "lsp_client"
+version = "0.1.0"
+edition = "2018"
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+lsp_json = { version = "0.1.0", path = "../lsp_json" }
+ng_async = { version = "0.1.0", path = "../ng_async" }
+emacs = { version = "0.1.0", path = "../emacs" }
+lisp-macros = { version = "0.1.0", path = "../lisp_macros" }
+lisp-util = { version = "0.1.0", path = "../lisp_util" }
+libc = "0.2.95"
+lazy_static = "1.2"
+lsp-server = "0.5.0"
+serde_json = { version = "1.0", features = ["preserve_order"] }
+
