+++ rust_src/crates/lsp_client/src/lib.rs
@@ -1,0 +1,16 @@
+#![allow(non_camel_case_types, non_snake_case, non_upper_case_globals)]
+#![feature(concat_idents)]
+#![feature(async_closure)]
+
+#[macro_use]
+extern crate serde_json;
+
+#[macro_use]
+extern crate emacs;
+#[macro_use]
+extern crate lisp_util;
+
+pub mod client;
+
+#[cfg(not(test))]
+include!(concat!(env!("CARGO_MANIFEST_DIR"), "/out/c_exports.rs"));
